requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
xrpl-py==2.6.0
pyTelegramBotAPI==4.17.0
cryptography==42.0.7
requests==2.32.3
gunicorn==22.0.0
matplotlib==3.9.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
requests-mock==1.11.0
